%pre client script: only run on upgrade, always exit 0
This script caused test failures, because it attempts to run on
fresh install but there is no `Requires(pre): alternatives` so
it runs before the alternatives binary is installed and fails.

It looks like it's only relevant on upgrade, so let's make the
script only run on upgrade, which should mean we don't need the
ordering dependency. And just for safety, we'll add `|| :` to
the command so it always exits 0 - this is required in the
packaging guidelines.

Signed-off-by: Adam Williamson <awilliam@redhat.com>