Make sure the /usr/sbin/lpc symlink is created on unmerged systems
The filetriggers in filesystem take care of creating compat symlinks
in /usr/sbin on unmerged systems. But /usr/bin/lpc is a %ghost file,
and the filetriggers are not triggered for it, so the symlink would
not be created (in some upgrade scenarios). Add it manually in the
scriptlet here.